Rename opentelemetry-auto-instrumentation to opentelemetry-instrumentation,and the console command `opentelemetry-auto-instrumentation` to `opentelemetry-instrument`
