+++ docs/auto_instrumentation/auto_instrumentation.rst
@@ -1,7 +1,7 @@
 OpenTelemetry Python Autoinstrumentation
 ========================================
 
-.. automodule:: opentelemetry.auto_instrumentation
+.. automodule:: opentelemetry.instrumentation
     :members:
     :undoc-members:
     :show-inheritance:
