+++ docs/auto_instrumentation/instrumentor.rst
@@ -1,7 +1,7 @@
-opentelemetry.auto_instrumentation.instrumentor package
+opentelemetry.instrumentation.instrumentor package
 =======================================================
 
-.. automodule:: opentelemetry.auto_instrumentation.instrumentor
+.. automodule:: opentelemetry.instrumentation.instrumentor
     :members:
     :undoc-members:
     :show-inheritance:
