+++ eachdist.ini
@@ -4,7 +4,7 @@
 sortfirst=
     opentelemetry-api
     opentelemetry-sdk
-    opentelemetry-auto-instrumentation
+    opentelemetry-instrumentation
     ext/opentelemetry-ext-wsgi
     ext/opentelemetry-ext-dbapi
     ext/*
