+++ ext/opentelemetry-ext-aiohttp-client/setup.cfg
@@ -40,6 +40,7 @@ package_dir=
 packages=find_namespace:
 install_requires =
     opentelemetry-api >= 0.9.dev0
+    opentelemetry-instrumentation == 0.9.dev0
     aiohttp ~= 3.0
 
 [options.packages.find]
