+++ ext/opentelemetry-ext-boto/setup.cfg
@@ -42,7 +42,7 @@ packages=find_namespace:
 install_requires =
     boto ~= 2.0
     opentelemetry-api == 0.9.dev0
-    opentelemetry-auto-instrumentation == 0.9.dev0
+    opentelemetry-instrumentation == 0.9.dev0
 
 [options.extras_require]
 test =
