+++ ext/opentelemetry-ext-dbapi/setup.cfg
@@ -41,6 +41,7 @@ package_dir=
 packages=find_namespace:
 install_requires =
     opentelemetry-api == 0.9.dev0
+    opentelemetry-instrumentation == 0.9.dev0
     wrapt >= 1.0.0, < 2.0.0
 
 [options.extras_require]
