+++ ext/opentelemetry-ext-redis/setup.cfg
@@ -41,7 +41,7 @@ package_dir=
 packages=find_namespace:
 install_requires =
     opentelemetry-api == 0.9.dev0
-    opentelemetry-auto-instrumentation == 0.9.dev0
+    opentelemetry-instrumentation == 0.9.dev0
     redis >= 2.6
     wrapt >= 1.12.1
