+++ ext/opentelemetry-ext-sqlite3/setup.cfg
@@ -42,7 +42,7 @@ packages=find_namespace:
 install_requires =
     opentelemetry-api == 0.9.dev0
     opentelemetry-ext-dbapi == 0.9.dev0
-    opentelemetry-auto-instrumentation == 0.9.dev0
+    opentelemetry-instrumentation == 0.9.dev0
     wrapt >= 1.0.0, < 2.0.0
 
 [options.extras_require]
