+++ opentelemetry-instrumentation/MANIFEST.in
