+++ opentelemetry-instrumentation/README.rst
@@ -1,0 +1,19 @@
+OpenTelemetry Auto Instrumentation
+==================================
+
+|pypi|
+
+.. |pypi| image:: https://badge.fury.io/py/opentelemetry-instrumentation.svg
+   :target: https://pypi.org/project/opentelemetry-instrumentation/
+
+Installation
+------------
+
+::
+
+    pip install opentelemetry-instrumentation
+
+References
+----------
+
+* `OpenTelemetry Project <https://opentelemetry.io/>`_
