+++ opentelemetry-instrumentation/setup.cfg
@@ -13,13 +13,13 @@
 # limitations under the License.
 #
 [metadata]
-name = opentelemetry-auto-instrumentation
+name = opentelemetry-instrumentation
 description = Auto Instrumentation for OpenTelemetry Python
 long_description = file: README.rst
 long_description_content_type = text/x-rst
 author = OpenTelemetry Authors
 author_email = cncf-opentelemetry-contributors@lists.cncf.io
-url = https://github.com/open-telemetry/opentelemetry-python/tree/master/opentelemetry-auto-instrumentation
+url = https://github.com/open-telemetry/opentelemetry-python/tree/master/opentelemetry-instrumentation
 platforms = any
 license = Apache-2.0
 classifiers =
@@ -48,5 +48,5 @@ where = src
 
 [options.entry_points]
 console_scripts =
-    opentelemetry-auto-instrumentation = opentelemetry.auto_instrumentation.auto_instrumentation:run
-    opentelemetry-bootstrap = opentelemetry.auto_instrumentation.bootstrap:run
+    opentelemetry-instrument = opentelemetry.instrumentation.auto_instrumentation:run
+    opentelemetry-bootstrap = opentelemetry.instrumentation.bootstrap:run
