+++ opentelemetry-instrumentation/setup.py
@@ -18,7 +18,7 @@ import setuptools
 
 BASE_DIR = os.path.dirname(__file__)
 VERSION_FILENAME = os.path.join(
-    BASE_DIR, "src", "opentelemetry", "auto_instrumentation", "version.py"
+    BASE_DIR, "src", "opentelemetry", "instrumentation", "version.py"
 )
 PACKAGE_INFO = {}
 with open(VERSION_FILENAME) as f:
